feat: Add Dockerfile and .dockerignore
The Dockerfile and .dockerignore files are added to the repository. The Dockerfile sets up the base image, installs necessary packages, and copies the application code. The .dockerignore file specifies the files and directories to be ignored when building the Docker image. These changes enable the application to be containerized and deployed using Docker.